fix: unbound extra_conf var in build_fftw.sh script
body: An unbound 'extra_conf' var was detected for the no_modules
case.  Due to the closing () surrounding the initialization of
the 'extra_conf' var on line 53 of build_fftw.sh, the variable
did not extend to the rest of the script and was limited to that
line.

The parenthases were removed and a default value of '' was included.