fix: revert some compile flag changes in hdf5/netcdf build scripts
body: Although supporting the generic linux build for gsi can
be achieved by setting hdf5 and netcdf as shared libraries, it
would be nice to have a unified noaa stack which also supports
jedi builds.  This is an attempt to compile generic linux version
of gsi without shared hdf5 and netcdf builds.